+++ Casks/menubar-ddc-control.rb
@@ -14,7 +14,7 @@ cask "menubar-ddc-control" do
   end
 
   depends_on arch: :arm64
-  depends_on macos: ">= :sonoma"
+  depends_on macos: :sonoma
 
   app "Menubar DDC Control.app"
   binary "#{appdir}/Menubar DDC Control.app/Contents/Resources/ddc-control"